Make sure we have pulled before reading LyraConfig
Move all initializing git pull into RepoGit and make RepoGit shared for
each unique local git clone.

As the repository specific configuration file is potentially stale,
we need to do this initializing pull before any read of that
configuration file. To ensure that, I replace all direct reads with a
read through RepoGit. Since the factory always pulls, we know the config
will be fresh after a Lyra server start.

Note that the route for creating pull-requests can now theoretically
pull twice, if RepoGit was not initialzied first. That is however
extremely unlikely as button the create pull requests is on a page that
uses RepoGit to load its data. Anyway, I beleive we will resolve this
when we implement resyncing from the remotes.