Update eslint-config-next to 13.5.6
eslint-config-next is built from the same source repository as next,
and they are released together.

When we set up the boilerplate, the intializer set up the dependencies
to use the same version. The versions only got out of sync later when
we copied eslint configuration from app.zetking.org in
e27f9cc5464ccc0a7c8a7eabe2c29d1869515f83

It is probably safer to keep next and eslint-config-next in sync.