Wrap constructor in factory
Instantiating objects with new can be a little ambigious. For example
new X().Y() is either creating a new instance with X as a constructor or
with Y as a constructor.

Being nameable, factory functions are also better than constructors when
you want to offer multiple different signatures for creating instances.