Hex encode value before logging
For this log line, the value is entirely controlled by
an unauthenticated client. Reflect that into logs allows
arbitrary log injection.

Hex encoding it is safer, at the expense of readability for
the administrator. This error case should be rare though so
sacrificing readability is not too expensive.

toHex will be moved to a better place
when we discover more usages for it.