Remove @emotion from direct dependencies
As far as I can tell, we do not use it directly. npm will install it
anyway as it is a peer dependency of our other dependencies.

Not declaring it in our package.json should give npm a little bit more
freedom in choosing compatible versions.

I tested this with npm run dev and the UI still looks good.