Refresh less
While its nice to have refresh built in to functions that benefit from
it, refresh is an expensive call and its important to not those
resources multiple times per request.

Lyra developers will have to be a little bit more careful to not forget
calling refresh appropriately. We can of course improve this futher
later to get both the performance and the nice developer experience,
but for now, this is a better trade-off.