Skip refresh if we did not fetch
If fetchAndCheckoutOriginBase was called so recently that we don't even
need to fetch again, then there are no new messages or translations to
read from file system and merge into project store.

Skipping refresh saves a lot of time in the page load.

Note that posting a new translation or creating a pull-request still
calls the refresh every time. These operations do not explicitly fetch
at all but a refresh might still be necessary if the project store has
not been initialized yet. Rather than think about this long enough,
I decided that it's ok that these operations are a little slow.